Enable sklearn test suites for sklearn_numba_dpex plugins
