+++ sklearn_numba_dpex/exceptions.py
@@ -1,0 +1,2 @@
+class NotImplementedError(NotImplementedError):
+    pass
